Add cross-validation script for threshold optimization
Implemented Fold 1 cross-validation for threshold determination. Threshold is determined based on the validation fold -- in this case, Fold1. Validation loss monitoring with a patience of 150 is used to track for overfitting. Early stopping when patience max is met. Then, validation AND training sets are tested. Thresholds calculated and printed to results.md for both.