Switch normalization from 99.9th percentile to mean/std.
Replaced 99.9th percentile normalization with mean and standard deviation-based normalization for NADH, FAD, and SHG channels. Now follows the formula xi = (xi - x̅ - 3sx)/6sx,

where xi is the individual image, x̅ is the dataset-wide mean of the optical parameter and sx is the masked dataset-wide standard deviation of the optical parameter (NADH, FAD, or SHG). This normalization method scales values within 3 standard deviations from the mean to between 0 and 1, effectively mapping the central portion of the data to the interval [0, 1]. Note that the subtraction and division operations in the equation are element-wise operations applied to each pixel of the image.